scraping_function/requirements.txt
requests
psycopg2-binary
elasticsearch==8.11.0
